// src/app/core/models/Session.ts
/**
 * @license
 * Copyright 2025 Google LLC
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
import {Event} from './types';

/**
 * Interface representing the state of a Session.
 */
export type SessionState = {
  [key: string]: unknown
};

/**
 * Interface representing a Session.
 */
export interface Session {
  id?: string;
  appName?: string;
  userId?: string;
  state?: SessionState;
  events?: Event[];
  lastUpdateTime?: number;
}